add authentication (PUT GET DELETE) for drink list

+++ drinks/views.py
@@ -3,7 +3,8 @@ from django.shortcuts import render
 from django.http import JsonResponse
 
 # rest framework
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, authentication_classes
+from rest_framework.authentication import TokenAuthentication
 from rest_framework.response import Response
 from rest_framework.exceptions import APIException
 from rest_framework import status
@@ -13,24 +14,28 @@ from .models import Drink
 from .serializers import DrinkSerializer
 
 
-# get all drinks - serialize them return json
-# accept get and post
 @api_view(['GET', 'POST'])
+@authentication_classes([TokenAuthentication])
 def drink_list(request):
-
     if request.method == 'GET':
         drinks = Drink.objects.all()
         serializer = DrinkSerializer(drinks, many=True)
         return Response(serializer.data)
 
-    serializer = DrinkSerializer(data=request.data)
-    serializer.is_valid(raise_exception=True)
-    # serializer.is_valid()
-    serializer.save()
-    return Response(serializer.data, status=status.HTTP_201_CREATED)
+    if request.user.is_authenticated:
+        request.data["createdBy"] = request.user.id
+        serializer = DrinkSerializer(data=request.data)
+        serializer.is_valid(raise_exception=True)
+        serializer.save()
+        return Response(serializer.data, status=status.HTTP_201_CREATED)
+
+    # ???
+    # how to give status code
+    raise APIException("Not authenticated.")
 
 
 @api_view(['GET', 'PUT', 'DELETE'])
+@authentication_classes([TokenAuthentication])
 def drink_detail(request, drink_id):
     try:
         drink = Drink.objects.get(pk=drink_id)
@@ -43,12 +48,17 @@ def drink_detail(request, drink_id):
             return Response(serializer.data, status=status.HTTP_201_CREATED)
     
         case 'PUT':
-            serializer = DrinkSerializer(drink, data=request.data)
-            # serializer.is_valid()
-            serializer.is_valid(raise_exception=True)
-            serializer.save()
-            return Response(serializer.data)
+            if request.user.is_authenticated and (request.user.username == drink.createdBy.username):
+                serializer = DrinkSerializer(drink, data=request.data)
+                serializer.is_valid(raise_exception=True)
+                serializer.save()
+                return Response(serializer.data)
+            # ???
+            # is this the correct form?
+            raise APIException("Not Authorized")
 
         case 'DELETE':
-            drink.delete()
-            return Response(status = status.HTTP_204_NO_CONTENT)
+            if request.user.is_authenticated and (request.user.username == drink.createdBy.username):
+                drink.delete()
+                return Response(status = status.HTTP_204_NO_CONTENT)
+            raise APIException("Not Authorized")
